Implement Fetching Of Post With GeoJSON data. Fixes T1482
Summary:
- Add use case for fetching post locally cached on the device
- Add use case for fetching post via the API
- Add post geo json entity
- Fix broken tests for the data module
- Add test for PostGeoJson entity mapper
- Fix main presenter broken test

Test Plan:
./gradlew test spoon

- See test pass

Reviewers: rjmackay

Subscribers: aMoniker

Maniphest Tasks: T1482

Differential Revision: https://phabricator.ushahidi.com/D817

+++ data/src/main/java/com/ushahidi/android/data/pref/LongPreference.java
@@ -64,7 +64,7 @@ public class LongPreference extends BasePreference<Long> {
     @Override
     public void set(Long value) {
         Preconditions.checkNotNull(value, "Long value cannot be null");
-        this.set(value);
+        this.set((long) value);
     }
 
     /**
